requirements.txt
tensorflow-cpu
Flask
Flask-RESTful
uwsgi
Flask-Cors==3.0.9
Flask-Uploads==0.2.1
Flask-WTF==0.14.3
scikit-image==0.18.1
gunicorn